Bug 1941552 - Fix Overflowed constant in nsMsgBiffManager::PerformBiff(). r=BenC
CID: 1639722

We do a i-- inside the loop. If that's at i = 0, next i is now 4294967295...

Differential Revision: https://phabricator.services.mozilla.com/D234289

--HG--
extra : rebase_source : 7c45e50b6e7244938a213b707208131a3213b2de
extra : amend_source : af61e48dc12de244171cce3edbf881b547a5765b

+++ mailnews/base/src/nsMsgBiffManager.cpp
@@ -274,7 +274,7 @@ nsresult nsMsgBiffManager::PerformBiff() {
   MOZ_LOG(MsgBiffLogModule, mozilla::LogLevel::Info, ("performing biffs"));
 
   uint32_t count = mBiffArray.Length();
-  for (uint32_t i = 0; i < count; i++) {
+  for (int32_t i = 0; i < (int32_t)count; i++) {
     // Take a copy of the entry rather than the a reference so that we can
     // remove and add if necessary, but keep the references and memory alive.
     nsBiffEntry current = mBiffArray[i];
